requirements.txt
telethon
python-dotenv
tabulate
flask